requirements.txt
google-adk>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai
